up ss3

+++ md4_b1/src/main/webapp/WEB-INF/views/list.jsp
@@ -18,7 +18,7 @@
     <table class="table table-bordered">
         <thead>
         <tr>
-            <th>ID</th>
+            <th>STT</th>
             <th>Mã cầu thủ</th>
             <th>Họ và tên</th>
             <th>Ngày sinh</th>
@@ -41,15 +41,37 @@
                     <img src="${player.avatarUrl}" alt="Ảnh đại diện" style="width: 100px; height: auto;">
                 </td>
                 <td>
-                    <a href="/player/${player.playerCode}/delete" class="btn btn-danger"
-                       onclick="return confirm('Bạn có chắc chắn muốn xóa cầu thủ này không?')">
-                        <i class="bi bi-trash"></i> Xóa
-                    </a>
+                    <button class="btn btn-danger" data-bs-toggle="modal" data-bs-target="#exampleModal${player.playerCode}">Remove</button>
+
+                    <div class="modal fade" id="exampleModal${player.playerCode}" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+                        <div class="modal-dialog">
+                            <div class="modal-content">
+                                <div class="modal-header">
+                                    <h5 class="modal-title" id="exampleModalLabel">Xóa cầu thủ</h5>
+                                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                                </div>
+                                <div class="modal-body">
+                                    Bạn có chắc chắn muốn xóa cầu thủ ${player.fullName}?
+                                    <div style="color: red; font-style: italic">
+                                        Hành động này không thể hoàn tác!
+                                    </div>
+                                </div>
+                                <div class="modal-footer">
+                                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
+                                    <a href="/player/${player.playerCode}/delete" class="btn btn-danger">Confirm</a>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
                 </td>
             </tr>
         </c:forEach>
         </tbody>
     </table>
 </div>
+
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
+        crossorigin="anonymous"></script>
 </body>
 </html>
